+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bug Fixes
 
 * Return `False` from `Additive.connected()` if the `_server` attribute is `None`
+* Assign id to simulations imported into a parametric study if the value is missing.
 
 ### Doc Improvements